reviewing and cleaning up code

+++ tweeter2_project/myapp/tweets.py
@@ -90,19 +90,45 @@ def tweets():
         conn = None
         own_post = True
         not_own_post = True
-        user_id = request.args.get('user_id')
+        user_id = request.args.get('userId')
 
         try:
             (conn, cursor) = dbConnect()
             if own_post:
                 cursor.execute("SELECT * FROM user INNER JOIN tweets on user.id = tweets.user_Id WHERE user_id=?", [user_id])
                 tweet_post = cursor.fetchall()
+                if cursor.rowcount > 0:
+                    user_tweets = []
+                    for tweet in tweet_post:
+                        tweets = {
+                            "tweetId": tweet[0],
+                            "userId": user_id,
+                            "username":tweet_post[0][1],
+                            "content": tweet[2],
+                            "createdAt":tweet[3]
+                        }
+                        user_tweets.append(tweets)
+                return Response(json.dumps(user_tweets, default=str),
+                            mimetype="application/json",
+                            status=201)
             elif not_own_post:
                 cursor.execute("SELECT * FROM user INNER JOIN tweet on user.id = tweets.user_Id")
                 tweet_post = cursor.fetchall()
-            return Response(json.dumps(tweet_post, default=str),
+                if cursor.rowcount > 0:
+                    all_tweets = []
+                    for tweet in tweet_post:
+                        tweets = {
+                            "tweetId": tweet[0],
+                            "userId": user_id,
+                            "username":tweet_post[0][1],
+                            "content": tweet[2],
+                            "createdAt":tweet[3]
+                        }
+                        all_tweets.append(tweets)
+                return Response(json.dumps(all_tweets, default=str),
                             mimetype="application/json",
                             status=201)
+           
     
         except ConnectionError:
             print("Error occured trying to connect to database")
@@ -134,22 +160,25 @@ def tweets():
         conn = None
         login_token = request.json.get('loginToken')
         content = request.json.get('content')
-        tweet_id = request.json.get('tweet_id')
+        tweet_id = request.json.get('tweetId')
 
         try:
             (conn, cursor) = dbConnect()
-            cursor.execute("SELECT user_id, FROM user INNER JOIN user_session ON user_session.user_id = user.id WHERE user_session.loginToken=?", [login_token,])
+            cursor.execute("SELECT user_id, loginToken FROM user_session INNER JOIN user ON user_session.user_id = user.id WHERE loginToken=?", [ login_token,])
             user = cursor.fetchall()
             user_id = user[0][0]
-            cursor.excute("UPDATE tweets SET content=? WHERE tweet_id=? AND user_id=?", [content, tweet_id, user_id])
-            conn.commit()
-            editedTweet = {
-                "tweetId" : tweet_id,
-                "content": content
-            }
-            return Response(json.dumps(editedTweet),
-                            mimetype="application/json",
-                            status=200)
+            cursor.execute("SELECT user_id FROM tweets WHERE id=?",[tweet_id,])
+            tweeter = cursor.fetchall()
+            if user[0][1] == login_token and user_id == tweeter[0][0]:
+                cursor.execute("UPDATE tweets SET content=? WHERE id=?", [content, tweet_id])
+                conn.commit()
+                editedTweet = {
+                    "tweetId" : tweet_id,
+                    "content": content
+                }
+                return Response(json.dumps(editedTweet),
+                                mimetype="application/json",
+                                status=200)
         
         except ConnectionError:
             print("Error occured trying to connect to database")
@@ -172,6 +201,7 @@ def tweets():
                 conn.close()
             else:
                 print("Failed to read data")
+
         return Response("Error something went wrong",
                         mimetype="text/plain",
                         status=500)
@@ -184,9 +214,12 @@ def tweets():
 
         try:
             (conn, cursor) = dbConnect()
-            cursor.execute("SELECT user_id, loginToken FROM user_session INNER JOIN user ON user_session.user_id = user.id")
-            user= cursor.fetchall()
-            if user[0][1] == login_token:
+            cursor.execute("SELECT user_id, loginToken FROM user_session INNER JOIN user ON user_session.user_id = user.id WHERE loginToken=?", [ login_token,])
+            user = cursor.fetchall()
+            user_id = user[0][0]
+            cursor.execute("SELECT user_id FROM tweets WHERE id=?",[tweet_id,])
+            tweeter = cursor.fetchall()
+            if user[0][1] == login_token and user_id == tweeter[0][0]:
                 cursor.execute("DELETE FROM tweets WHERE id=?", [tweet_id])
                 conn.commit()
                 return Response("Tweet deleted",
